# API/storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Project Ecommerce API",
        "description": "Tài liệu API cho hệ thống Ecommerce",
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://localhost",
            "description": "Local server"
        }
    ],
    "paths": {
...
        "/api/v1/auth/signup": {
            "post": {
                "tags": [
                    "Auth"
                ],
                "summary": "Đăng ký tài khoản mới",
                "operationId": "49e6d1fe647ef7467c487fccc631b0e7",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "name",
                                    "email",
                                    "password"
                                ],
                                "properties": {
                                    "name": {
                                        "type": "string",
                                        "example": "Nguyen Van A"
                                    },
                                    "email": {
                                        "type": "string",
                                        "format": "email",
                                        "example": "user@example.com"
                                    },
                                    "password": {
                                        "type": "string",
                                        "format": "password",
                                        "example": "12345678"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Đăng ký thành công, vui lòng xác thực email!"
                    },
                    "422": {
                        "description": "Dữ liệu không hợp lệ"
                    }
                }
            }
        },
        "/api/v1/auth/login": {
            "post": {
                "tags": [
                    "Auth"
                ],
                "summary": "Đăng nhập",
                "operationId": "24404833453b6de86bc74a8b91fee5ab",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "email",
                                    "password"
                                ],
                                "properties": {
                                    "email": {
                                        "type": "string",
                                        "format": "email",
                                        "example": "user@example.com"
                                    },
                                    "password": {
                                        "type": "string",
                                        "format": "password",
                                        "example": "12345678"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Login successful!"
                    },
                    "401": {
                        "description": "Sai thông tin đăng nhập"
                    }
                }
            }
        },
        "/api/v1/auth/logout": {
            "post": {
                "tags": [
                    "Auth"
                ],
                "summary": "Đăng xuất",
                "operationId": "485ea771a9e5bec87095353a38457063",
                "responses": {
                    "200": {
                        "description": "Logout successful!"
                    },
                    "401": {
                        "description": "Logout failed"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/v1/auth/tokens/refresh": {
            "post": {
                "tags": [
                    "Auth"
                ],
                "summary": "Làm mới token",
                "operationId": "274183ebbafa7660f9bd651636e464e7",
                "responses": {
                    "200": {
                        "description": "Refresh token successful!"
                    },
                    "401": {
                        "description": "Refresh token failed. Please try again!"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/v1/password/forgot": {
            "post": {
                "tags": [
                    "Auth"
                ],
                "summary": "Gửi email quên mật khẩu",
                "operationId": "3a949e640b9b03e921ab372335a9ea2a",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "email"
                                ],
                                "properties": {
                                    "email": {
                                        "type": "string",
                                        "format": "email",
                                        "example": "user@example.com"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Password reset link sent successfully!"
                    },
                    "500": {
                        "description": "Failed to send password reset link."
                    }
                }
            }
        },
        "/api/v1/password/reset": {
            "post": {
                "tags": [
                    "Auth"
                ],
                "summary": "Đặt lại mật khẩu",
                "operationId": "1fd5a993e43ee5b122c2f199f697002a",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "email",
                                    "password",
                                    "password_confirmation",
                                    "token"
                                ],
                                "properties": {
                                    "email": {
                                        "type": "string",
                                        "format": "email",
                                        "example": "user@example.com"
                                    },
                                    "password": {
                                        "type": "string",
                                        "format": "password",
                                        "example": "12345678"
                                    },
                                    "password_confirmation": {
                                        "type": "string",
                                        "format": "password",
                                        "example": "12345678"
                                    },
                                    "token": {
                                        "type": "string",
                                        "example": "reset-token"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Password reset successfully!"
                    },
                    "500": {
                        "description": "Failed to reset password."
                    }
                }
            }
        },
...
    },
    "components": {
...
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    },
...
}